# (Fix) Redis async context usage for Redis Cluster + 94% lower median latency when using Redis Cluster  