## (Fix) Redis async context usage

## Issues this PR Solves
- Redis Cluster implementation was raising `coroutine ignored GeneratorExit` at around 100 RPS 


# Load Test Results 

Locust Test was 100 New Users, 20 user Ramp Up calling a fake endpoint

## Redis Cluster Load Test
Metric | Before PR | With PR | Improvement
-- | -- | -- | --
Median Latency | 120ms | 7ms | 94% faster with this PR
99th Percentile | 300ms | 22ms | 93% faster with this PR


## Redis Node Load Test

(This is quite small almost negligible)

Metric | Before PR | With PR | Improvement
-- | -- | -- | --
Median Latency | 94ms | 92ms | 2% faster
99th Percentile | 130ms | 110ms | 15% faster







<br class="Apple-interchange-newline">

<!-- e.g. "Implement user authentication feature" -->

## Relevant issues

<!-- e.g. "Fixes #000" -->

## Type

<!-- Select the type of Pull Request -->
<!-- Keep only the necessary ones -->

🐛 Bug Fix
🧹 Refactoring
🚄 Infrastructure
✅ Test

## Changes

<!-- List of changes -->

## [REQUIRED] Testing - Attach a screenshot of any new tests passing locally
If UI changes, send a screenshot/GIF of working UI fixes

<!-- Test procedure -->